Move PEP related functions and command to own cog
